Predict clinical and molecular features

Main function used to predict various clinical and molecular features from gene-level methylation data in prostate cancer patients.

Usage

estimate.features(
  gene.methy.data,
  models,
  prop.missing.cutoff = 0.3,
  validate.data = TRUE
)

Arguments

gene.methy.data

A data frame with gene-level methylation data, created by gene.methylation. Patients are rows and columns are genes.

models

A list of models used to predict features from gene-level methylation data. The models should come from data('all.models').

prop.missing.cutoff

The maximum proportion of missing values allowed for each required gene.

validate.data

TRUE/FALSE, whether to validate input data. This should generally always be TRUE, but developers may set to FALSE to speedup testing/development.

Value

A list with the following objects:

features a data frame with predicted features as columns and patients as rows

validation results from validation checks telling you which features you were able to predict. See validate.gene.methy.data for more details.

Examples

data('all.models');

### example gene-level methylation data
data('example.data.gene.methy');
# note this dataset is derived from the following commands:
# data('example.data');
# example.data.gene.methy <- gene.methylation(example.data);

features <- estimate.features(example.data.gene.methy, all.models);
#> [1] "Starting imputation..."
#> [1] "Finished imputation."
str(features);
#> List of 2
#>  $ features  :'data.frame':	10 obs. of  14 variables:
#>   ..$ age.continuous      : num [1:10] 52.6 59.6 61.2 57 56.3 ...
#>   ..$ ISUP.grade          : Factor w/ 5 levels "ISUP1","ISUP2",..: 2 3 2 2 2 2 5 1 2 2
#>   ..$ t.category          : Factor w/ 4 levels "T1","T2","T3",..: 2 4 2 3 3 2 3 3 2 3
#>   ..$ psa.categorical     : Factor w/ 3 levels "psa_lt_10","psa_10_19.9",..: 1 1 1 1 1 1 1 1 1 1
#>   ..$ pga                 : num [1:10] 10.16 10.58 8.55 28.04 7.23 ...
#>   ..$ CHD1.cna.loss       : Factor w/ 2 levels "Yes","No": 2 2 2 2 2 2 2 2 2 2
#>   ..$ NKX3.1.cna.loss     : Factor w/ 2 levels "Yes","No": 2 1 1 1 1 2 2 1 2 2
#>   ..$ MYC.cna.gain        : Factor w/ 2 levels "Yes","No": 2 2 2 1 2 2 1 1 2 2
#>   ..$ PTEN.cna.loss       : Factor w/ 2 levels "Yes","No": 2 1 2 1 1 2 2 2 2 2
#>   ..$ CDKN1B.cna.loss     : Factor w/ 2 levels "Yes","No": 2 2 2 2 2 2 1 2 2 2
#>   ..$ RB1.cna.loss        : Factor w/ 2 levels "Yes","No": 2 2 1 2 2 1 2 2 1 2
#>   ..$ CDH1.cna.loss       : Factor w/ 2 levels "Yes","No": 2 2 2 2 2 2 2 2 2 2
#>   ..$ TP53.cna.loss       : Factor w/ 2 levels "Yes","No": 2 2 2 2 2 2 2 2 2 2
#>   ..$ log2p1.snvs.per.mbps: num [1:10] 0.709 0.994 0.857 1.073 0.862 ...
#>  $ validation:List of 5
#>   ..$ val.passed                      : logi TRUE
#>   ..$ features.you.can.predict        : Named logi [1:14] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   .. ..- attr(*, "names")= chr [1:14] "age.continuous" "ISUP.grade" "t.category" "psa.categorical" ...
#>   ..$ required.genes                  :List of 14
#>   .. ..$ age.continuous      : chr [1:13603] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ ISUP.grade          : chr [1:5000] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ t.category          : chr [1:5000] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ psa.categorical     : chr [1:100] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ pga                 : chr [1:5000] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ CHD1.cna.loss       : chr [1:1000] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ NKX3-1.cna.loss     : chr [1:13603] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ MYC.cna.gain        : chr [1:13603] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ PTEN.cna.loss       : chr [1:13603] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ CDKN1B.cna.loss     : chr [1:5000] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ RB1.cna.loss        : chr [1:1000] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ CDH1.cna.loss       : chr [1:13603] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ TP53.cna.loss       : chr [1:13603] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   .. ..$ log2p1.snvs.per.mbps: chr [1:5000] "NIPA2" "CNBP" "DDX55" "MRPS25" ...
#>   ..$ missing.genes                   :List of 14
#>   .. ..$ age.continuous      : chr(0) 
#>   .. ..$ ISUP.grade          : chr(0) 
#>   .. ..$ t.category          : chr(0) 
#>   .. ..$ psa.categorical     : chr(0) 
#>   .. ..$ pga                 : chr(0) 
#>   .. ..$ CHD1.cna.loss       : chr(0) 
#>   .. ..$ NKX3-1.cna.loss     : chr(0) 
#>   .. ..$ MYC.cna.gain        : chr(0) 
#>   .. ..$ PTEN.cna.loss       : chr(0) 
#>   .. ..$ CDKN1B.cna.loss     : chr(0) 
#>   .. ..$ RB1.cna.loss        : chr(0) 
#>   .. ..$ CDH1.cna.loss       : chr(0) 
#>   .. ..$ TP53.cna.loss       : chr(0) 
#>   .. ..$ log2p1.snvs.per.mbps: chr(0) 
#>   ..$ required.genes.with.high.missing:List of 14
#>   .. ..$ age.continuous      : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ ISUP.grade          : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ t.category          : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ psa.categorical     : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ pga                 : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ CHD1.cna.loss       : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ NKX3-1.cna.loss     : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ MYC.cna.gain        : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ PTEN.cna.loss       : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ CDKN1B.cna.loss     : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ RB1.cna.loss        : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ CDH1.cna.loss       : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ TP53.cna.loss       : Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0) 
#>   .. ..$ log2p1.snvs.per.mbps: Named num(0) 
#>   .. .. ..- attr(*, "names")= chr(0)