Predict methylation subtype

Assign patients to four prostate cancer DNA methylation subtypes

Usage

estimate.subtypes(
  methy.data,
  prop.missing.cutoff = 0.3,
  impute.using.all.cpgs = TRUE
)

Arguments

methy.data

A data.frame with patients as rows (rownames give patient ids) and column names give CpG ids.

prop.missing.cutoff

The maximum proportion of missing values allowed for each required CpG. KNN imputation is used to impute missing values.

impute.using.all.cpgs

TRUE/FALSE indicating whether to impute missing values using all CpGs in methy.data or only the CpGs required by subtype.model. When TRUE, imputation will be slower and use more memory, but should be more accurate.

Examples

data('subtype.model');

### example CpG data
data('example.data');

subtypes <- estimate.subtypes(example.data);
#> [1] "Starting imputation..."
#> Warning: 5611 rows with more than 50 % entries missing;
#>  mean imputation used for these rows
#> [1] "Finished imputation."
head(subtypes);
#>              subtype
#> TCGA-HC-8213    MS-3
#> TCGA-G9-6363    MS-1
#> TCGA-G9-6343    MS-4
#> TCGA-KK-A8II    MS-3
#> TCGA-FC-7708    MS-1
#> TCGA-EJ-A65G    MS-1